[Refactor] 변경된 예외 처리 파일들의 경로를 적용 - 변경된 예외 처리 파일들의 경로를 적용
