+++ backend/ricecake/src/main/java/project/ricecake/member/service/MemberService.java
@@ -3,9 +3,9 @@ package project.ricecake.member.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import project.ricecake.common.BaseResponse;
-import project.ricecake.error.exception.UserDuplicateException;
-import project.ricecake.error.exception.UserNotFoundException;
-import project.ricecake.error.exception.InvalidPasswordException;
+import project.ricecake.error.exception.duplicate.UserDuplicateException;
+import project.ricecake.error.exception.notfound.UserNotFoundException;
+import project.ricecake.error.exception.unauthorized.InvalidPasswordException;
 import project.ricecake.member.domain.request.PostLoginReq;
 import project.ricecake.member.domain.request.PostSignupReq;
 import project.ricecake.member.domain.entity.MemberEntity;
